requirements.txt
# Core Game Dependencies
pygame>=2.6.0           # Graphics, audio, and input handling
pydantic>=2.0.0         # Data validation and models
PyYAML>=6.0             # YAML content loading
watchdog>=3.0.0         # Hot-reload file monitoring

# AI Art Generation Dependencies
diffusers>=0.29.0       # Stable Diffusion and other models
transformers>=4.35.0    # Model loading and inference
torch==2.3.0+cu121      # PyTorch with CUDA 12.1 support
xformers>=0.0.22        # Memory-efficient attention for Transformers

# Image Processing
Pillow>=10.0.0          # Image manipulation and processing
opencv-python>=4.8.0    # Computer vision and image processing
numpy>=1.24.0           # Numerical operations

# Development and Testing
pytest>=7.4.0           # Testing framework
pytest-asyncio>=0.21.0  # Async test support
pytest-cov>=4.1.0       # Coverage reporting
black>=23.7.0           # Code formatting
flake8>=6.0.0           # Code linting
mypy>=1.5.0             # Type checking

# Utilities
requests>=2.31.0        # HTTP requests for ComfyUI integration
psutil>=5.9.0           # System monitoring for performance tests
click>=8.1.0            # Command-line interface utilities

# Optional Dependencies (for specific features)
# Uncomment as needed:

# Audio Processing (for advanced audio features)
# librosa>=0.10.0        # Audio analysis
# soundfile>=0.12.0      # Audio file I/O

# Advanced Graphics (for shader effects)
# moderngl>=5.8.0        # OpenGL wrapper
# PyOpenGL>=3.1.0        # OpenGL bindings

# Machine Learning (for AI opponents)
# scikit-learn>=1.3.0    # Machine learning algorithms
# tensorflow>=2.13.0     # Deep learning (alternative to PyTorch)

# Networking (for multiplayer features)
# websockets>=11.0       # WebSocket support
# asyncio-mqtt>=0.13.0   # MQTT for real-time communication

# Packaging and Distribution
# pyinstaller>=5.13.0    # Executable building
# cx-Freeze>=6.15.0      # Alternative executable builder

# Platform-Specific Dependencies
# Windows-specific (uncomment on Windows)
# pywin32>=306           # Windows API access

# macOS-specific (uncomment on macOS)  
# pyobjc>=9.2            # macOS API access

# Linux-specific (uncomment on Linux)
# python-xlib>=0.33      # X11 library access